# web-v2/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_PaginatedData_PlanSummary__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_PlanSummary_"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiEnvelope[PaginatedData[PlanSummary]]",
        "type": "object"
      },
      "ApiEnvelope_PaginatedData_RunHeader__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_RunHeader_"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiEnvelope[PaginatedData[RunHeader]]",
        "type": "object"
      },
...
      "ApiEnvelope_PlanDetailData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanDetailData"
              },
              {
                "type": "null"
              }
            ]
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiEnvelope[PlanDetailData]",
        "type": "object"
      },
...
      "ApiFailureEnvelope": {
        "additionalProperties": false,
        "description": "Failure-only envelope used for declared error responses.",
        "properties": {
          "data": {
            "title": "Data",
            "type": "null"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "minItems": 1,
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "data",
          "errors"
        ],
        "title": "ApiFailureEnvelope",
        "type": "object"
      },
...
      "PaginatedData_RunHeader_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunHeader"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "PaginatedData[RunHeader]",
        "type": "object"
      },
...
      "PlanStatus": {
        "description": "Supported Plan statuses.",
        "enum": [
          "ready",
          "applying",
          "applied",
          "partial_failed",
          "failed",
          "cancelled",
          "expired"
        ],
        "title": "PlanStatus",
        "type": "string"
      },
...
      "PlanType": {
        "description": "Supported Plan types.",
        "enum": [
          "add",
          "organize",
          "refresh",
          "undo"
        ],
        "title": "PlanType",
        "type": "string"
      },
...
      "RunStatus": {
        "description": "Supported Run statuses.",
        "enum": [
          "running",
          "succeeded",
          "partial_failed",
          "failed"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "OMYM2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/history": {
      "get": {
        "operationId": "getHistory",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "plan_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plan Id"
            }
          },
          {
            "in": "query",
            "name": "library_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Library Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_PaginatedData_RunHeader__"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get History"
      }
    },
...
    "/api/plans": {
      "get": {
        "operationId": "listPlans",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PlanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PlanType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "blocked",
            "required": false,
            "schema": {
              "default": false,
              "title": "Blocked",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_PaginatedData_PlanSummary__"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Plans"
      }
    },
    "/api/plans/{plan_id}": {
      "get": {
        "operationId": "getPlan",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_PlanDetailData_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailureEnvelope"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "X-OMYM2-Correlation-ID": {
                "description": "Request correlation identifier written to server logs.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Plan"
      }
    },
...
  }
}